app/requirements.txt
attrs==17.4.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
configparser==3.5.0
cx-Freeze==5.1.1
cycler==0.10.0
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
flake8==3.5.0
Flask==0.12.2
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-Sessionstore==0.4.5
Flask-SocketIO==2.9.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
funcsigs==1.0.2
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
greenlet==0.4.13
gunicorn==19.1.0
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
mosspy==1.0.5
oauth2==1.9.0.post1
pathlib==1.0.1
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pycode-similar==1.2
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.4.2
pytest==3.3.2
python-dateutil==2.6.1
python-engineio==2.0.4
python-google-places==1.4.1
python-socketio==1.9.0
requests==2.18.4
shutilwhich==1.1.0
six==1.11.0
SQLAlchemy==1.2.4
urllib3==1.22
virtualenv-clone==0.2.6
Werkzeug==0.14.1
whitenoise==3.3.1
yelp==1.0.2
